% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_site_core.R
\name{dm_site_core}
\alias{dm_site_core}
\title{Generate Table of Randomized Participants by Site}
\usage{
dm_site_core(
  population_from,
  population_where = NULL,
  idvar = "USUBJID",
  treatment_var = "TRT01P",
  treatment_order = NULL,
  site_var = "SITEID",
  country_var = NULL,
  level_col_name = FALSE,
  display_total_list = NULL
)
}
\arguments{
\item{population_from}{Subject-level input dataset, typically ADSL.}

\item{population_where}{A character vector of the filtering condition applied to the subject-level dataset.}

\item{idvar}{A string representing the variable name for subject ID.}

\item{treatment_var}{A string representing the variable name for the treatment.}

\item{treatment_order}{A character vector of treatment groups in the desired order.}

\item{site_var}{A string representing the variable name for the site, default is 'SITEID'.}

\item{country_var}{A string representing the variable name for the country, default is NULL.
This is optional; if not provided, country information will be excluded.}

\item{level_col_name}{A logical parameter indicating whether to include the leveling columns.}

\item{display_total_list}{A list of named lists for aggregating treatment groups.}
}
\value{
A list containing two data frames: one named report.table.n for count results,
and the other named report.table.p for percentage results.
}
\description{
Generates two summary data frames grouped by treatment and site and
optionally by country. Both data frames share the same structure, with one
containing counts and the other percentages. The function takes ADSL as
the input dataset.

For each country and/or site, it summarizes the number of unique subjects
and their corresponding percentages in each treatment group, presenting
the results as a country/site × treatment matrix. In the count data frame,
the top row displays the total number of subjects in each treatment group,
while in the percentage data frame, these values are represented as NAs.
}
\examples{
dm_site_core(
  population_from = demo_adsl,
  population_where = "FASFL == 'Y'",
  idvar = "USUBJID",
  treatment_var = "TRT01P",
  treatment_order = NULL,
  site_var = "SITEID",
  country_var = "COUNTRY",
  level_col_name = TRUE,
  display_total_list = NULL
 )
}
