% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitStats.R
\name{fitStats}
\alias{fitStats}
\title{Calculate model statistical summary}
\usage{
fitStats(actual, predict)
}
\arguments{
\item{actual}{real value response variable}

\item{predict}{predicted response variable}
}
\value{
mape, rmse, mae
}
\description{
Function to calculate model summary, include mean absolute percentage error,
root-mean-square error, mean absolute error
}
