% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getMX}
\alias{getMX}
\title{Get annual maxima from data.table}
\usage{
getMX(dta, a = 1)
}
\arguments{
\item{dta}{data.table including the fields date (datum) and p (precipitation in kg/s)}

\item{a}{the number of intervals to aggregate over in hours}
}
\value{
data.table with year, aggregation and maxima
}
\description{
Get annual maxima from data.table
}
\examples{
\dontrun{
getMX(dta, a = c(1, 3))
}
}
