% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunctions.R
\name{h}
\alias{h}
\title{A short function for head, plot, names}
\usage{
h(x, ...)
}
\arguments{
\item{data}{data object, vec or df to pass}
}
\description{
This function replace View with see
}
\examples{
h(c(1,2,3))
p(x, y)
n(dataset)
}
\keyword{calls}
\keyword{fast}
\keyword{library}
\keyword{to}
