% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunctions.R
\name{lib}
\alias{lib}
\title{A Function to load l}
\usage{
lib(lib_tocheck)
}
\arguments{
\item{library}{string parameter to pass}
}
\description{
This function allows load libraries and if needed import them on the fly
}
\examples{
lib("ggplot2")
}
\keyword{lib}
\keyword{library}
