% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pha}
\alias{pha}
\title{SpatialPolygonsDataFrame representing the Prague area}
\format{
An object of class \code{SpatialPolygonsDataFrame} with 1 rows and 8 columns.
}
\usage{
pha
}
\description{
SpatialPolygonsDataFrame representing the Prague area
}
\keyword{datasets}
