% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{precip_ALADIN_hist}
\alias{precip_ALADIN_hist}
\title{Simulated hourly precipitation}
\format{
data.table with
\describe{
  \item{data}{date and time}
  \item{p}{precipitation in kg/s}
  \item{forcing}{forcing string id}
}
}
\usage{
precip_ALADIN_hist
}
\description{
The simulation was done with ALADIN RCM within the CORDEX project
}
\keyword{datasets}
