% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printDbname.R
\name{printDbname}
\alias{printDbname}
\title{printDbname}
\usage{
printDbname(client_id, user, password)
}
\arguments{
\item{client_id}{client account id}

\item{user}{user name}

\item{password}{usr password}
}
\description{
print data base name
}
