% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryDbname.R
\name{queryDbname}
\alias{queryDbname}
\title{query database name for clients}
\usage{
queryDbname(client_id, user, password)
}
\arguments{
\item{client_id}{account id}

\item{user}{user account id}

\item{password}{user password}
}
\description{
Function to query databse name for specific client
}
