% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunctions.R
\name{see}
\alias{see}
\title{A short function for View}
\usage{
see(dat)
}
\arguments{
\item{data}{data object, vec or df to pass}
}
\description{
This function replace View with see
}
\examples{
see(c(1,2,3))
}
\keyword{calls}
\keyword{fast}
\keyword{library}
\keyword{to}
