\name{toy-package}
\alias{toy-package}
\alias{toy}
\docType{package}
\title{
\packageTitle{toy}
}
\description{
\packageDescription{toy}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{toy}
\packageIndices{toy}
 An overview of how to use the package, including the most important functions 
}
\author{
\packageAuthor{toy}

Maintainer: \packageMaintainer{toy}
}
\references{
 Literature or other references for background information 
}
\keyword{ package }
\seealso{
Optional links to other man pages, e.g. 
\code{\link[<pkg>:<pkg>-package]{<pkg>}}
}
\examples{
simple examples of the most important functions
}
