% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy-package.R
\docType{package}
\name{toy-package}
\alias{toy}
\alias{toy-package}
\title{toy: A Toy R Package for Reporting Tables in Safety Statistics}
\description{
Automates the creation of reporting tables for safety statistics using standardized input datasets like ADaMs. Includes tables for demographics, disposition, adverse reactions, AE summaries, specific AEs, and more.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/yilu-jin/toy}
  \item Report bugs at \url{https://github.com/yilu-jin/toy/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
